modified:   pom.xml renamed:    src/main/java/banje/ban/GamePanel_chap1.java -> src/main/java/chapter/chap1/GamePanel_chap1.java renamed:    src/main/java/banje/ban/Rock.java -> src/main/java/chapter/chap1/Rock_chap1.java renamed:    src/main/java/banje/ban/GamePanel_chap2.java -> src/main/java/chapter/chap3/GamePanel_chap3.java renamed:    src/main/java/banje/ban/Rock_chap2.java -> src/main/java/chapter/chap3/Rock_chap3.java renamed:    src/main/java/banje/ban/MAIN.java -> src/main/java/game/MAIN.java deleted:    target/classes/banje/ban/GamePanel_chap2$1.class deleted:    target/classes/banje/ban/MAIN.class renamed:    target/classes/banje/ban/GamePanel_chap1.class -> target/classes/chapter/chap1/GamePanel_chap1.class renamed:    target/classes/banje/ban/Rock.class -> target/classes/chapter/chap1/Rock_chap1.class new file:   target/classes/chapter/chap3/GamePanel_chap3$1.class renamed:    target/classes/banje/ban/GamePanel_chap2.class -> target/classes/chapter/chap3/GamePanel_chap3.class renamed:    target/classes/banje/ban/Rock_chap2.class -> target/classes/chapter/chap3/Rock_chap3.class new file:   target/classes/game/MAIN.class
