+++ src/main/java/chapter/chap3/Rock_chap3.java
@@ -1,29 +1,33 @@
-package banje.ban;
+package chapter.chap3;
 
 import java.awt.Color;
 import java.awt.Graphics;
 import java.awt.Rectangle;
 import java.awt.event.KeyEvent;
 
-public class Rock {
-    public final int ScreenHeight =750;
-    public final int ScreenWidth = 750;
+public class Rock_chap3 {
+    final int TypeSize = 12;
+    final int scale = 10;
+    final int tileSize = TypeSize * scale;
+    final int maxScreenCol = 8;
+    final int maxScreenRow = 7;
+    final int ScreenWidth = tileSize * maxScreenCol;
+    final int ScreenHeight = tileSize * maxScreenRow;   
     public int x; 
     public int y;  
     public int preX = x, preY = y;
-    private final int tileSize = 25;
     Rectangle rect;
-    public Rock(int x, int y) {
-        this.x = x;
-        this.y = y;
+    public Rock_chap3(int x, int y) {
+        this.x = x * tileSize;
+        this.y = y * tileSize;
         this.preX = this.x; // Khởi tạo preX
         this.preY = this.y; // Khởi tạo preY
         rect = new Rectangle(this.x, this.y, tileSize, tileSize); // Khởi tạo Rectangle
     }
     // Phương thức để vẽ nhân vật Rock
     public void draw(Graphics g) {
-        g.setColor(Color.red); 
-        g.fillOval(x, y, 50 , 50 );  // Vẽ hình tròn tại vị trí (x, y)
+        g.setColor(Color.white);  // Đặt màu trắng
+        g.fillOval(x, y, tileSize, tileSize);  // Vẽ hình tròn tại vị trí (x, y)
     }
     public boolean vaCham(Rectangle rectA) {
         if (rectA.intersects(rect)) return true;
